Issue #75, Added seperate files for the Temperature board.
Also moved chip specific code into seperate directory, Library/SPI.
This will make it possible to use the same SW for temperature boards
version 4.3 and below. Note, this version works for 4.3 and below.
