+++ STM32/Temperature/.cproject
@@ -114,6 +114,7 @@
                                     <listOptionValue builtIn="false" value="&quot;${workspace_loc:/Temperature/Libraries/SI7051/Inc}&quot;"/>
                                     									
                                     <listOptionValue builtIn="false" value="&quot;${workspace_loc:/Temperature/Libraries/Util/Inc}&quot;"/>
+				    <listOptionValue builtIn="false" value="&quot;${workspace_loc:/Temperature/Libraries/SPI/Inc}&quot;"/>
                                     								
                                 </option>
                                 								
@@ -172,8 +173,8 @@
                         <entry flags="VALUE_WORKSPACE_PATH|RESOLVED" kind="sourcePath" name="Drivers"/>
                         						
                         <entry flags="VALUE_WORKSPACE_PATH" kind="sourcePath" name="Libraries/SI7051"/>
-                        						
                         <entry flags="VALUE_WORKSPACE_PATH" kind="sourcePath" name="Libraries/USBprint"/>
+			<entry flags="VALUE_WORKSPACE_PATH" kind="sourcePath" name="Libraries/SPI"/>
                         						
                         <entry flags="VALUE_WORKSPACE_PATH" kind="sourcePath" name="Libraries/Util"/>
