+++ STM32/Temperature/Core/Inc/Temperature.h
@@ -1,0 +1,6 @@
+#pragma once
+
+#include "stm32f4xx_hal.h"
+
+void InitTemperature(SPI_HandleTypeDef* hspi_, I2C_HandleTypeDef* hi2c_);
+void LoopTemperature();
